polarisation consolidation transformation
