+++ src/ibex_bluesky_core/devices/simpledae/__init__.py
@@ -24,7 +24,6 @@ from ibex_bluesky_core.devices.simpledae._reducers import (
     PeriodGoodFramesNormalizer,
     PeriodSpecIntegralsReducer,
     ScalarNormalizer,
-    polarization,
     sum_spectra,
     tof_bounded_spectra,
     wavelength_bounded_spectra,
@@ -71,7 +70,6 @@ __all__ = [
     "Waiter",
     "check_dae_strategies",
     "monitor_normalising_dae",
-    "polarization",
     "sum_spectra",
     "tof_bounded_spectra",
     "wavelength_bounded_spectra",
